users@infra has no dominion here
jira approvals and rejections must be handled by the project.
if there is a technical issue, {project} can open an infra jirat.
if there is a technical issue, {project} can open an infra jirat.
if there is a technical issue, {project} can open an infra jirat.
if there is a technical issue, {project} can open an infra jira.